% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattk_colours.R
\docType{data}
\name{mattk_colours}
\alias{mattk_colours}
\title{Colours used by the package}
\format{
An object of class \code{list} of length 11.
}
\usage{
mattk_colours
}
\description{
Colours used by the package
}
\keyword{datasets}
